+++ qiskit/quantum_info/__init__.py
@@ -5,11 +5,13 @@
 # This source code is licensed under the Apache License, Version 2.0 found in
 # the LICENSE.txt file in the root directory of this source tree.
 
-"""Quantum Information  methods."""
+"""Quantum Information methods."""
 
+from .operators.operator import Operator
 from .operators.unitary import Unitary
 from .operators.pauli import Pauli, pauli_group
 from .operators.channel import Choi, SuperOp, Kraus, Stinespring, Chi, PTM
-from .states.states import basis_state, random_state, projector, purity
-from .states.measures import state_fidelity
 from .operators.measures import process_fidelity
+from .states.states import basis_state, projector, purity
+from .states.measures import state_fidelity
+from .random import random_unitary, random_state, random_density_matrix